+++ classes/profilefields.php
@@ -240,7 +240,7 @@ abstract class profilefields {
 
     /**
      * Generate tabs for the display
-     * @return \tabtree
+     * @return \core\output\tabtree
      */
     protected function get_tabs() {
         $tabs = [];
@@ -250,7 +250,7 @@ abstract class profilefields {
                                  get_string('addrules', 'local_profilecohort'));
         $tabs = array_merge($tabs, $this->extra_tabs());
 
-        $tabtree = new \tabtree($tabs, $this->action);
+        $tabtree = new \core\output\tabtree($tabs, $this->action);
 
         return $tabtree;
     }