+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2025-04-14 - Upgrade: Update Bootstrap accordion to Bootstrap 5.0 on "cohort members" page
 * 2025-04-14 - Prepare compatibility for Moodle 5.0.
 
 ### v4.5-r1